graphGenerator/requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0.0